+++ folioflex/portfolio/assets.py
@@ -85,7 +85,7 @@ def update_asset_info(
             params = config_helper.get_config_options(
                 config_path, "assets", "houses", item
             )
-            value = wrappers.Zillow().get_value(params, proxy=proxy)
+            value = wrappers.Zillow().get_value(params)
         else:
             logger.error(f"Asset group '{asset_group}' not found.")
             return None
@@ -207,16 +207,19 @@ def get_checking_value(engine: database.Engine, user: Optional[str] = None) -> f
         left_on="item_id",
         right_on="id",
         how="left",
-        suffixes=[None, "_tmp"],
+        suffixes=[None, "_item"],
     )
+    account_df.drop(columns=["id_item"], inplace=True)
+
     account_df = pd.merge(
         account_df,
         user_df[["id", "username"]],
         left_on="user_id",
         right_on="id",
         how="left",
-        suffixes=[None, "_tmp"],
+        suffixes=[None, "_user"],
     )
+    account_df.drop(columns=["id_user"], inplace=True)
     if user is not None:
         account_df = account_df[account_df["username"] == user]
     account_df = account_df[account_df["subtype"] == "checking"]